requirements.txt
Flask==2.2.5
numpy==1.24.4
Pillow==10.0.0
joblib==1.3.1
